[android] Cleaned up documentation; fixed build
